Make the spy action choice dialog work again
Freeciv patch #4476 (svn commit 24375) made the server ask what spy action
should be taken against a tile in stead of asking what spy action to take
against (one of) the potential victim(s) on the tile. The question was also
moved to a separate packet.

Modify the web client so it handles this.