Allow stealing a tech from a victim without establishing an embassy first
Move the check that the victim has a technology to steal from the steal tech
spy action to the server. On the client it would fail when the owner of the
diplomat didn't have access to a list of the victim's techs. That list
requires an embassy to the victim.

See bug #21718