Rule set: Disable unit bribery in Democracies via "Bribe Unit"
The Unbribable_Units effect has been removed in Freeciv. Acived the same by
using the action enabler of the "Bribe Unit" action in stead. This way there
won't be a need to do it when the Freeciv version is updated. As a bonus
people won't be fooled by an enabled action button when requesting actions
vs a unit from a Democracy.