protocol: stop messing with unit_do_action
Leave the type of unit_do_action's value field alone. The Freeciv server is
able to handle SINT16 fields in the JSON protocol. On top of that the field
is never used by Freeciv-web. It is used by targeted sabotage and targeted
steal tech to sent the ID of the target. Freeciv-web doesn't support those
spy actions yet.