Show server suggested city names with quote chars
The server can suggest a city name containing the single quote character.
The suggested city name used to be added in an unescaped form to a
JavaScript generated HTML string. The apostrophe in the city name would then
be understood as "quote over" by the browser executing the HTML code.

Add the suggested city name to the text input by setting its value attribute
from JavaScript. This fixes the issue in the case of accepting city name
suggestions.

Note that a city name with an apostrophe may cause the same problem in other
areas of the Freeciv-web client JavaScript code.

Reported by Andreas Røsdal <andreasr>

See bug #23835