Switch "Detonate Nuke" to "Detonate Nuke At"
Having a nuke blow up inside the city where it is stationed because of a
misunderstanding is a nasty surprise. Change "Detonate Nuke" to the same
model the other Freeciv clients use. When the order is given a goto appears.
The nuke goes to the tile the user sends it to and detonates there.