webperimental: city founding on foreign terrain
Allow cities to be founded within foreign borders during war and before
contact. The idea is that founding a city inside someone's terrain is an act
of war.

Don't allow founding cities inside the terrain of allies and team mates.
This is to prevent a small poxer allied to a large poxers from destroying
the large poxer by founding small cities between the large poxer's cities.