pregame settings: allow changing back to fcweb.
The the pregame settings dialog's ruleset setting will only change the
current ruleset if it has changed. It assumed that the current ruleset was
fcweb. This wasn't always correct.

It therefore became impossible to use the settings dialog to change from
webperimental to fcweb. Changing it by manually using the '/ruleset'
command worked fine.

Set the ruleset setting to the current ruleset. This makes it notice a
change from webperimental to fcweb.