web client: support action decision want.
The Freeciv server now signals that it would be a good idea to pop up the
action selection dialog for a unit via fields in the unit data. Remove the
patches that reverted this change. Support it in the web client.

Note that the Freeciv-web client won't benefit from having a wanted action
decision in the unit data. The Freeciv-web client never had the diplomat
queue of the C clients. It opens multiple action selection dialogs behind
each other. The server will drop the action decision data when it replies
with the actions the unit can do. This auto dropping is gone in upstream
versions of the Freeciv server Freeciv-web haven't reached.