Remove the servlet's db password from Git.
The configuration file context.xml stores the database user name and the
password for the Java servlet. Freeciv-web won't work if the Java servlet
don't have the correct password. Changing the password in context.xml can
result in it accidentally being included in a commit.

Move the version controlled context.xml to context.xml.dist. Have it copied
to context.xml. This makes it easier to change the password.