Stop assuming no stored action selection dialogs.
Have the client ask the server to mark that a unit wants to know what
actions it can perform to its current tile rather than asking it directly
what actions can be performed. This costs an extra round trip. In return the
code won't break in a hard to detect way if Freeciv-web gets some form of
action selection queue.