Quick fix unit multi select problems.
The implementation of right click and drag to select multiple units could
break the Freeciv-web user interface in some cases. The problem is that the
context menu is opened on right click mouse down. The context menu then goes
on to prevent the mouseup event from firing when the right mouse button is
released. This breaks right click to recenter. It also prevents the unit
multi selection from finishing. The last problem leaves the unit selection
box on the player map until the player presses the escape button.

Make right mouse button down block the context menu from popping up. Right
mouse button up removes the context menu block. The escape button does the
same.

Actively show the context menu when a right mouse button up is seen as a
request for the context menu.

This fixes the right click to multi select problems I was able to find in
Chromium Version 51.0.2704.79 Built on 8.4, running on Debian 8.5 (64-bit)
and in Firefox 45.2.0 on Debian Stable. No testing has been performed on
mobile devices.

This fix should probably also be tested on other browsers, especially those
the old code worked on, before release. The patch it self should also be
reviewed by someone with a decent knowledge of the Freeciv-web UI code
before release.