Web client: ask for target for "Do...".
Change "Do..." to always ask the player for a target tile. A 2nd "Do..."
(like pressing "d" twice) will specify the tile the actor is on.

This gives the player a way to select the actor unit's target tile. A
ruleset can allow a unit without any movement left to act against a neighbor
tile. It also allow supporting the actions "Paradrop Unit" and
"Airlift Unit" in the action selection dialog.

Pressing "d" twice doesn't slow down acting against a target on the same
tile as the actor too much. No mouse is required.

A user with an existing habit of pressing "d" to act against (a target on)
the tile the actor is likely to press it again if the first key press don't
work. If he selected the menu item he will notice that the mouse pointer
turns into a selector.