Update to Freeciv server r33631.
webperimental: set the dispersion server setting to 2 by default.

Dispersing the start units, together with patch #7543 and patch #7587, gives
a Webperimental game a distinct visual first impression. The different look
of the game start helps to communicate that the Webperimental rules really
are different from the Classic rules.

See patch #7606