update-wikipedia-docs.py: get government docs.
Add the government names to the list of items to get Wikipedia articles for.

This commit doesn't regenerate freeciv-wiki-doc.js with the government names
added. The fact that freeciv-wiki-doc.js is version controlled makes me
suspect that updates to it may follow some procedure I'm not familiar with.
(Like checking for Wikipedia prank edits.)