Provision packages for new boxes
When bringing up a new box after having used another one you may get a "Failed to start nginx.service: Unit nginx.service not found." error. Tomcat, maven, JDK, etc. are also not installed.

vagrant-build.sh uses /vagrant/freeciv-web/target/freeciv-web.war as a kind of canary: when it doesn't exist, it gets the needed packages and whatever else it does. If it exists, it means freeciv-web has been built and deployed before, so it thinks everything is ready and it just starts freeciv-web.

This patch adds /usr/sbin/nginx to the test, although the right way would probably be to split the script in three parts at least: provisioning, building and starting freeciv-web, and then check and run each one when it's necessary.