pregame: viewable full ruleset description.
Rulesets like civ2civ3 has a short description shown when loaded and a full
description in a README file. The full description is shown in the manual's
"About Current Ruleset" section.

A player picking the ruleset of a game using the C client has the README
file at his computer. A Freeciv-web player don't. The manual can't be read
before the game is started.

Make the full ruleset description viewable in the pregame. Put it next to
the ruleset chooser.