Keep Cities tab order between updates.
tablesorter doesn't seem to have a way to query for current order. But it marks the header cells, so we use that. It allows to sort by several columns; if that is used, we don't know the column priorities, so in that case the result may not be the same that the user had.