Remove hardcoded z-index for generic modal dialog.
At least in firefox, when the dialog is closed automatically with the timeout while having a browser dialog or menu open, subsequent calls make the z-index of the "modal layer overlay" have a z-index of 2001, 2002, etc., with the setting of that of the dialog to 2000 making it inaccessible (and darkened).

It also seems unnecessary, as not setting it keeps the stacking correct.

If needed, should prefer .dialog('moveToTop').