Don't close (and steal text input focus) a closed dialog.
When a general dialog is closed manually (escape, close button or Ok button), the timeout to close it should be cancelled to avoid "closing it again" (which may affect a _different_ dialog and stealing focus from the chat box.

That timeout seems to be the root cause of the unexpected loss of keyboard focus in the chat box reported in <http://forum.freeciv.org/f/viewtopic.php?f=24&t=75573&p=81412&hilit=focus#p81408>. Having seen it, now it's easy to reproduce: middle-click a tile to get its info, close the dialog, go to the chat box and start typing. 24 seconds after having queried the tile info, focus is relinquished by the general keyboard handler. Another effect is that if you get the info for several tiles in less than those 24 seconds, the last dialog may be closed too soon, affected by the timeout of a previous dialog.

Note that I'm keeping the timeout, just cancelling it when the dialog is closed, because I don't know what it is needed for (that "because sometimes the dialog can't be closed manually" is both too general and too scary). If it's not needed anymore, ignore this patch and remove the timeout.