Handle multi and no selection in nations list.
When the nations tab is opened, with no selected nation, only the game scores button should be enabled, if at all.

The selectable widget allows multiple selection, be it by ctrl-click or by dragging the mouse. Currently we are keeping just one player as selected internally, so make that the same in the UI. And as you can unselect, update the buttons and the internal data in that case.

Also, there's no nation_button class anymore.

Handling multiple selection (instead of avoiding it) might be worthwhile for the withdraw vision and cancel treaty actions.

And the oneliner to "disable" multiple selection is not perfect. When selecting by dragging the mouse from bottom to top, the selectable widget reports the events in DOM order, so the item that is selected at the end of the interaction is the one in which it started, not the one where it finished. If won't allow multiple selection we should probably ditch selectable and just add a click handler (and maybe something for the keyboard?).