Get back Year, /help and /show messages
The check for longturn chat may be a bit overzealous, and it filters
some system messages also. This lets the year change messages and
the response from /help and /show commands to pass through.
If all the command responses are safe and have the same format,
those two regexes can be changed to /^<font [^>]*>\/[^: ]*: /.