Scroll nations list in touch enabled devices
The selectable plugin is very much geared towards multiple selection,
with even a lasso widget for that. That means a drag is used for
selection, and not for scroll or any other thing, there's a
touch-action: none in ui-selectable items. That seems to disable
scrolling in touch-enabled devices.

We don't need multiple selection in the nations list (because we don't
currently do anything with it, although it might be marginally useful
for some of the actions), and simple click handling does the job.