Tweak nations list buttons enablement
There's no point in enabling some actions when the selected player is
the one playing or is dead. There are still some conditions that are
not checked (for example the meet player action may be active even if
the other player is an AI and the ruleset doesn't allow diplomacy
with AIs).

Also, let observers get full intel from players.