Close city dialog only once
When calling close_city_dialog from any place but the dialog's cross
button, there's a call to close the dialog, which in turn calls
close_city_dialog again. The second call to close_city_dialog is
unnecessary, and the call to close the dialog is only needed from
outside the close handler, so break that in two: close_city_dialog
now just asks the dialog to close, whereas the rest of the stuff is
executed in the handler.