Fix: ws is null from setup_hotseat_game
setup_hotseat_game is called after a timeout, but the websocket may not still exist (it is created in websocket_init after a server response to a POST request).

Same for create_pbem_players, although much more difficult in this case due to the higher timeout.

Reported by Track.js.