Use common install script for docker
The specific tomcat image (or any other) does not play well with our
install/update stuff, as that's not the way docker is supposed to be
used. So we change to a basic system image.

The end result is a multigigabyte system with everything.

TODO: small, specific, directed containers with just the deployed
services, and not the development tools, source and history.