Support 'graphic_alt' for buildings in web.
A ruleset can add an alternative graphics tag to a building (aka improvement)
type in case the tileset don't have the sprite specified by the 'graphic'
ruleset field ('graphic_str' in the network protocol). This makes it possible
to add a new building to a ruleset that will prefer a custom building image
but will fall back to a building image present in all tilesets. An example of
a ruleset where this is done is "Alien World".

Support using the 'graphic_alt' tag for buildings in Freeciv-web.