Update to Freeciv server 3fd93791e8
Freeciv commit message:
Fix network protocol for rulesets with lots of units.

There was a longstanding bug if a ruleset has 129 or more units,
and one of the 129th-or-later units is an upgrade or convert target,
the client wouldn't register that, so players couldn't initiate the
affected upgrade or conversion. This was due to use of SINT8 in the
network protocol.

Change the network protocol definition to UINT8. It so happens that
this doesn't actually change what the server sends on the wire, just
how the client interprets it, so there's no need for a capability
bump, and new clients will now do the correct thing even with old
servers.

Reported by Akechi.

See hrm bug #763894.