Fix: use extras list to show resources
The resource field only makes sense in the server, it is used to know
a tile has a resource even if terrain changes and it's not valid, in
case it changes back. It's not even rebuilt for the private map when
loading a game. Real valid resources are in the extras list.

Fixes #41 (resources not shown from a loaded savegame)
Fixes fish on swamp and other such resources in changed terrain