Remove tabs keydown handler from main tabs
The tabs widget adds keydown handlers to tabs and content to navigate them with the keyboard. This is not good in the map, as some of those keys are used for units movement and the handlers stop propagation of the event. Until now, the issue has been avoided adding activation and click handlers that remove focus from tabs and content so they don't get keyboard events. We'll need the map to handle keyboard events, so let's remove the plugin handlers instead. And it's better to remove what we don't need than to add a workaround.