Support ruleset change in PBEM
There's no UI for ruleset selection in PBEM games. The player can set it through the chat line, though, but will not notice that phasemode has been changed and will probably start a PBEM game in concurrent mode, which won't advance much. The same problem exists in hotseat games, despite having the UI to change ruleset.

The player can and should `/set phasemode player` after changing ruleset, but it'd be better if they didn't have to, so let's do it for them.