Fix: Check auth kind in proxy
freeciv-proxy is the entry point to a running game. It does check auth info with the server tables, but:
- The call to Google auth is client-controlled
- Sending the response to the TokenSignin servlet is client-controlled
- The auth data passed to the freeciv-proxy is client-controlled
- The checks in freeciv-proxy are lax, to say the least

So visitors could, with just a bit of js run in their browser, log in to a otpd game without having signed in to Google and even without an account in the freeciv-web server.