Goto path request should not modify unit state
This fixes a longstanding, annoying issue where the server could report
units as `done_moving` at the beginning of a turn. This prevented any
such units from 1) being cycled through using wait (W) and 2) would
falsely enable the "Turn Done" button when units still had remaining
moves.

The cause of the issue was that selecting a unit with the mouse sends
a path request to the server, which modified the servers state for
the unit. The modifications made would indirectly cause the server
to mark the unit as `done_moving` until the unit was given other
orders. The issue could be triggered by viewing goto paths for a unit
without selecting any path to actually goto.

Viewing path advice should not modify state of a unit, so removing
modifications made in the goto patch fixes the error.