Update script.lua
this should fix an error Andreas discovered... it turns out this was the wrong file and was script.lua from multiplayer rules instead of mpplus.  the error was it was trying to find darwin's voyage to give free tech, but mpplus got rid of darwin's voyage and renamed it "voyage of darwin" so it does not give free tech.