Fix: run apt-get in noninteractive mode
Some updates may ask questions to the user, but running under Vagrant
and other unattended provisioners there may not be a way to do so.

DEBIAN_FRONTEND was already set as "noninteractive", but sudo rightfully
slashes the environment and passes only a handful of variables, so apt-get
was not seeing that value.

Closes #285