Update to Freeciv server freeciv/freeciv@0adbbc46
This is just one commit forward. It brings in "Transport Embark"
action. As web-client is not smart enough to automatically know
when it should request embark action, this causes sligh UI
degradation. User can't simply move unit to transport, but
needs to select "Embark" from the action selection dialog that
pops up.

Signed-off-by: Marko Lindqvist <cazfi74@gmail.com>