Introduce urgent unit focus.
Allow units to jump to the front of the unit focus queue.

Being able to jump to the front of the unit focus queue is strongly
recommended to implement an action selection dialog that follows the unit
focus queue system like the C clients have had since Freeciv 2.6. Without it
the probability that a unit is killed while waiting for input increases.