Tolerate ending unstarted action selection.
Relax the assertion in action_selection_no_longer_in_progress() so it allows
ending the action selection for a unit that never started it. This may
happen when a follow up question to the action selection dialog is asked
from a non action selection dialog source like a menu item for that action
alone.

See osdn #42042