Prepare for server repository revisions with Sentry Unit action
1) Update server commit to fetch to freeciv/freeciv@fc6101ecfd
2) Backport patch to remove Sentry Unit action

The new commit from the server repository adds Sentry Unit action.
It was present in the repository for some time only. We don't want
to have it in freeciv-web even temporarily (to switch back and forth).
So we backport removal of that Sentry Unit action as a patch.
Now we can update server forward with the patch to remove the action
in place. It had to be modified a bit to apply to this server
revision, so it will require modifications (back to original form) as
server gets updated to revisions closer to one where it was
backported from.

Signed-off-by: Marko Lindqvist <cazfi74@gmail.com>