Update to Freeciv server freeciv/freeciv@b3b07c597b
Freeciv commit message:
Enabler control over regular moves.

Introduce the new action "Unit Move" making regular unit movement action
enabler controlled.

A consequence is that code for giving regular moves the same properties as
if it was controlled by the action system can go away. Included in this
patch is the replacement of the action selection dialogue's "Keep moving"
button, the removal of the custom handling for actions.move_is_blocked_by
and the removal of various "do a move action or regular move"-code. Clean
ups of other areas aren't included in this patch.

The optimizations required to make it possible to move the movement cost to
the ruleset are not included, neither is the move of that cost itself.

See hrm Feature #919970