Update to Freeciv server freeciv/freeciv@d946395bf4
Freeciv commit message:
Add the new DiplRelTile requirement type.

It is like the DiplRel requirement type except that it is about the owner of
the tile.

This allows rules like "a unit can only be expelled if it is on a tile owned
by someone its owner has peace with".

See osdn #42515