Update to Freeciv server freeciv/freeciv@62ed22545d
Freeciv commit message:
Ruleset actor consuming always for "Found City".

This allows a ruleset where you get a unit that can found cities at the
start of the game but are unable to build a city founding unit. (Think of a
GamleLoss unbuildable "King" unit as the only city founder.)

See hrm Feature #931180