Dockerfile: Give user running server access to the scenario files
I'm not really happy with this quick solution, but it's one
that made the scenario loading to work at all.

Implemented in the Dockerfile level, to not mess with other
kind of setups. If the issue is not specific to the Docker
setup, we should implemented it in more generic part of
the install.

Signed-off-by: Marko Lindqvist <cazfi74@gmail.com>