Show correct score
Show player scores on nations dialog as network protocol
tells it. Any positive score is shown, negative values
(indicating that the client does not know the score) are not.

This fixes both cases where the web-client showed negative
values as real score, and failed to show the score when
the user is supposed to know it.

Signed-off-by: Marko Lindqvist <cazfi74@gmail.com>